requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
progressbar2==3.53.1
python-utils==2.5.6
requests==2.25.1
six==1.16.0
urllib3==1.26.4
xeger==0.3.5
yapf==0.31.0
